requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pdf2image==1.16.3
Pillow==9.5.0
PyPDF2==3.0.1
pytesseract==0.3.10
